Update Columns Shortcode
The column shortcodes are a little stiff and are not allowing the full flexibility of the column system. I'd like to the option to make the columns both the first and the last in the row so a full 12 span column can be used. Extra extendability for collapsing the column would useful too.